lethe-research/experiments/requirements.txt
# Lethe Research Experiment Dependencies
# Core MLflow integration for Phase 2.4

# MLflow for experiment tracking and reproducibility
mlflow>=2.10.0
mlflow[extras]>=2.10.0  # Includes additional dependencies for various backends

# Core scientific computing and ML
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep learning and NLP
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Data processing and storage
pyyaml>=6.0
jsonlines>=3.1.0
# sqlite3 - Built into Python
psutil>=5.9.0

# Statistical analysis
statsmodels>=0.14.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Async and HTTP
aiofiles>=23.0.0
httpx>=0.25.0
requests>=2.31.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0

# FastAPI service dependencies (for prediction service integration)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0